Refactor Gallery
