+++ src/pages/index.js
@@ -1,8 +1,15 @@
-import FilterPanel from "components/FilterPanel";
-import PostGallery from "components/PostGallery";
+import { useState } from "react";
 import Layout from "components/Layout";
+import FilterPanel from "components/FilterPanel";
+import SortControls from "components/SortControls";
+import Gallery from "components/Gallery";
+import Pagination from "@material-ui/lab/Pagination";
 
 export default function Home() {
+  const [sort, setSort] = useState(sortItems[0]);
+  const handleSort = (e) => {
+    setSort(e.target.name);
+  };
   return (
     <Layout
       title="iFLEXHIBIT"
@@ -10,7 +17,19 @@ export default function Home() {
     >
       <div className="home-container">
         <FilterPanel tags={tags} />
-        <PostGallery />
+        <div className="posts-container">
+          <SortControls
+            items={sortItems}
+            sortBy={sort}
+            handleSort={handleSort}
+          />
+          <Gallery posts={posts} grid={3} />
+          <Pagination
+            count={10}
+            color="primary"
+            style={{ display: "flex", justifyContent: "center" }}
+          />
+        </div>
       </div>
     </Layout>
   );
@@ -34,4 +53,47 @@ const tags = [
   "Traditional Art",
   "Wallpaper",
   "Video",
+];
+
+const sortItems = [
+  "NEW TODAY",
+  "NEW THIS WEEK",
+  "POPULAR (14 DAYS)",
+  "POPULAR (ALL TIME)",
+];
+
+const posts = [
+  {
+    author: {
+      avatar: "/img/boi.jpg",
+      username: "sosig69",
+    },
+    likes: 69,
+    views: 1000000,
+    comments: 420,
+    title: "Cyberpunk 2069 - Kirk Sawyer",
+    image: "/img/cards/card4.png",
+  },
+  {
+    author: {
+      avatar: "/img/boi.jpg",
+      username: "sosig69",
+    },
+    likes: 69,
+    views: 1000000,
+    comments: 420,
+    title: "Cyberpunk 2069 - Cloud Girl",
+    image: "/img/cards/card2.png",
+  },
+  {
+    author: {
+      avatar: "/img/boi.jpg",
+      username: "sosig69",
+    },
+    likes: 69,
+    views: 1000000,
+    comments: 420,
+    title: "Cyberpunk 2069 - Kerry Eurodyne",
+    image: "/img/cards/card3.png",
+  },
 ];
