+++ src/styles/Gallery.scss
@@ -1,25 +1,37 @@
-div.profile-gallery {
-  padding: 1rem 2rem;
-  display: flex;
-  flex-flow: column nowrap;
-  gap: 1rem;
-
-  & > &-header {
-    display: grid;
-    grid-template-columns: auto 0.5fr;
-    align-items: flex-end;
+div.gallery- {
+  &2 {
+    grid-template-columns: repeat(2, 1fr);
+    & > div.card {
+      height: 20rem;
+      div.card-info-group {
+        padding: 0.5rem 1rem;
+      }
+    }
   }
 
-  & > div.gallery {
+  &3 {
+    grid-template-columns: repeat(3, 1fr);
+    & > div.card {
+      height: 15rem;
+      div.card-info-group {
+        padding: 0 0.5rem 0.5rem;
+      }
+    }
+  }
+
+  &2,
+  &3 {
+    margin: 1rem 0;
     display: grid;
-    grid-template-columns: repeat(2, 1fr);
     gap: 0.25rem;
 
     & > div.card {
       position: relative;
+
       width: 100%;
-      height: 20rem;
+
       border-radius: 5px;
+      background-color: var(--black);
       overflow: hidden;
 
       &:hover {
@@ -54,16 +66,15 @@ div.profile-gallery {
 
         & > div.card-info-group {
           height: 30%;
-          padding: 1rem;
 
           display: flex;
           flex-flow: column nowrap;
-          align-items: flex-start;
           justify-content: space-between;
           overflow: hidden;
 
           & > div.card-info-group-item {
             width: 100%;
+            height: 100%;
 
             display: flex;
             flex-flow: row nowrap;
@@ -91,6 +102,7 @@ div.profile-gallery {
 
           &:last-of-type {
             & > div.card-info-group-item {
+              justify-content: flex-end;
               & > span {
                 font-family: var(--serif);
               }
