--- src/styles/PostGallery.scss
@@ -1,100 +1,0 @@
-div.posts {
-  & > div.gallery {
-    margin: 1rem 0;
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 0.25rem;
-
-    & > div.card {
-      position: relative;
-
-      width: 100%;
-      height: 15rem;
-
-      border-radius: 5px;
-      background-color: var(--black);
-      overflow: hidden;
-
-      &:hover {
-        & > div.card-info {
-          opacity: 1;
-        }
-      }
-
-      & > div.card-info {
-        z-index: 1;
-
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        bottom: 0;
-
-        border-radius: 5px;
-        background: linear-gradient(var(--blue-fade) 40%, var(--black-75) 85%);
-        color: var(--white);
-
-        font-size: 0.875rem;
-        font-weight: 700;
-
-        display: flex;
-        flex-flow: row nowrap;
-        align-items: flex-end;
-        justify-content: space-between;
-
-        opacity: 0;
-        transition: 125ms ease-in-out;
-
-        & > div.card-info-group {
-          height: 30%;
-          padding: 0.5rem;
-
-          display: flex;
-          flex-flow: column nowrap;
-          align-items: flex-start;
-          justify-content: space-between;
-          overflow: hidden;
-
-          & > div.card-info-group-item {
-            width: 100%;
-
-            display: flex;
-            flex-flow: row nowrap;
-            align-items: center;
-            justify-content: space-between;
-
-            gap: 0.5rem;
-          }
-
-          &:first-of-type {
-            & > div.card-info-group-item {
-              justify-content: flex-start;
-              & > div.avatar {
-                position: relative;
-                width: 1.5rem;
-                height: 1.5rem;
-                border-radius: 5px;
-                overflow: hidden;
-              }
-              & > span {
-                white-space: nowrap;
-              }
-            }
-          }
-
-          &:last-of-type {
-            & > div.card-info-group-item {
-              & > span {
-                font-family: var(--serif);
-              }
-              & > i {
-                opacity: 0.75;
-                color: var(--white);
-              }
-            }
-          }
-        }
-      }
-    }
-  }
-}
