+++ src/styles/ProfileBio.scss
@@ -1,4 +1,4 @@
-div.profile-details {
+div.profile-bio {
   width: 100%;
   padding: 2rem;
   display: flex;
