+++ src/styles/containers.scss
@@ -17,3 +17,7 @@ div.profile-body-container {
   display: grid;
   grid-template-columns: 0.3fr auto;
 }
+
+div.posts-container {
+  margin: 2rem;
+}
