+++ src/styles/globals.scss
@@ -56,7 +56,7 @@ main {
 
 @import "containers.scss";
 @import "SortControls.scss";
-@import "PostGallery.scss";
+@import "Gallery.scss";
 @import "FilterPanel.scss";
 @import "CheckList.scss";
 @import "TextField.scss";
@@ -64,6 +64,5 @@ main {
 @import "Navbar.scss";
 @import "SiteFooter.scss";
 @import "ProfileBanner.scss";
-@import "ProfileDetails.scss";
-@import "ProfileGallery.scss";
+@import "ProfileBio.scss";
 @import "media.scss";
